feat: Update the useAuthentication custom hook to use isLoading state
The isLoading state in the useAuthentication hook is necessary to handle
the loading state of the authentication check. When we check if the user
is logged in by sending a request to the server, it takes some time to
get a response. During this time, the authentication
status is unknown yet.

By using the isLoading state, we can keep track of whether the
authentication check is still in progress or has completed. Initially,
isLoading is set to true to indicate that the check is ongoing. Once we
receive the response and determine the authentication status, isLoading
is set to false to show that the check is finished.